+++ tests/Console/ModelsCommand/Pivot/__snapshots__/Test__test__1.php
@@ -5,17 +5,26 @@ declare(strict_types=1);
 namespace Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Pivot\Models;
 
 use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Pivot\Models\Pivots\CustomPivot;
+use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Pivot\Models\Pivots\DifferentCustomPivot;
 use Illuminate\Database\Eloquent\Model;
 
 /**
  * 
  *
- * @property-read CustomPivot $customAccessor
+ * @property-read DifferentCustomPivot|CustomPivot|null $pivot
+ * @property-read \Illuminate\Database\Eloquent\Collection<int, ModelWithPivot> $relationCustomPivotUsingSameAccessor
+ * @property-read int|null $relation_custom_pivot_using_same_accessor_count
+ * @property-read CustomPivot|null $customAccessor
  * @property-read \Illuminate\Database\Eloquent\Collection<int, ModelWithPivot> $relationWithCustomPivot
  * @property-read int|null $relation_with_custom_pivot_count
- * @method static \Illuminate\Database\Eloquent\Builder|ModelWithPivot newModelQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|ModelWithPivot newQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|ModelWithPivot query()
+ * @property-read DifferentCustomPivot|null $differentCustomAccessor
+ * @property-read \Illuminate\Database\Eloquent\Collection<int, ModelWithPivot> $relationWithDifferentCustomPivot
+ * @property-read int|null $relation_with_different_custom_pivot_count
+ * @property-read \Illuminate\Database\Eloquent\Collection<int, ModelWithPivot> $relationWithDifferentCustomPivotUsingSameAccessor
+ * @property-read int|null $relation_with_different_custom_pivot_using_same_accessor_count
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|ModelWithPivot newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|ModelWithPivot newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|ModelWithPivot query()
  * @mixin \Eloquent
  */
 class ModelWithPivot extends Model
@@ -25,6 +34,28 @@ class ModelWithPivot extends Model
         return $this->belongsToMany(ModelwithPivot::class)
             ->using(CustomPivot::class)
             ->as('customAccessor');
+    }
+
+
+    public function relationWithDifferentCustomPivot()
+    {
+        return $this->belongsToMany(ModelwithPivot::class)
+            ->using(DifferentCustomPivot::class)
+            ->as('differentCustomAccessor');
+    }
+
+    // without an accessor
+
+    public function relationCustomPivotUsingSameAccessor()
+    {
+        return $this->belongsToMany(ModelwithPivot::class)
+            ->using(CustomPivot::class);
+    }
+
+    public function relationWithDifferentCustomPivotUsingSameAccessor()
+    {
+        return $this->belongsToMany(ModelwithPivot::class)
+            ->using(DifferentCustomPivot::class);
     }
 }
 <?php
@@ -38,11 +69,30 @@ use Illuminate\Database\Eloquent\Relations\Pivot;
 /**
  * 
  *
- * @method static \Illuminate\Database\Eloquent\Builder|CustomPivot newModelQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|CustomPivot newQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|CustomPivot query()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CustomPivot newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CustomPivot newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CustomPivot query()
  * @mixin \Eloquent
  */
 class CustomPivot extends Pivot
+{
+}
+<?php
+
+declare(strict_types=1);
+
+namespace Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Pivot\Models\Pivots;
+
+use Illuminate\Database\Eloquent\Relations\Pivot;
+
+/**
+ * 
+ *
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|DifferentCustomPivot newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|DifferentCustomPivot newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|DifferentCustomPivot query()
+ * @mixin \Eloquent
+ */
+class DifferentCustomPivot extends Pivot
 {
 }