+++ tests/Console/ModelsCommand/SoftDeletes/__snapshots__/Test__test__1.php
@@ -11,13 +11,13 @@ use Illuminate\Database\Eloquent\SoftDeletes;
  * 
  *
  * @property int $id
- * @method static \Illuminate\Database\Eloquent\Builder|Simple newModelQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|Simple newQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|Simple onlyTrashed()
- * @method static \Illuminate\Database\Eloquent\Builder|Simple query()
- * @method static \Illuminate\Database\Eloquent\Builder|Simple whereId($value)
- * @method static \Illuminate\Database\Eloquent\Builder|Simple withTrashed()
- * @method static \Illuminate\Database\Eloquent\Builder|Simple withoutTrashed()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|Simple newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|Simple newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|Simple onlyTrashed()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|Simple query()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|Simple whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|Simple withTrashed()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|Simple withoutTrashed()
  * @mixin \Eloquent
  */
 class Simple extends Model