+++ tests/Console/ModelsCommand/UnionTypes/__snapshots__/Test__test__1.php
@@ -14,11 +14,11 @@ use Illuminate\Database\Query\Builder;
  * @property-read string|int|null $foo
  * @property-read \Illuminate\Database\Eloquent\Collection<int, UnionTypeModel> $withUnionTypeReturn
  * @property-read int|null $with_union_type_return_count
- * @method static \Illuminate\Database\Eloquent\Builder|UnionTypeModel newModelQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|UnionTypeModel newQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|UnionTypeModel query()
- * @method static \Illuminate\Database\Eloquent\Builder|UnionTypeModel withNullableUnionTypeParameter(string|int|null $bar)
- * @method static \Illuminate\Database\Eloquent\Builder|UnionTypeModel withUnionTypeParameter(string|int $bar)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|UnionTypeModel newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|UnionTypeModel newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|UnionTypeModel query()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|UnionTypeModel withNullableUnionTypeParameter(string|int|null $bar)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|UnionTypeModel withUnionTypeParameter(string|int $bar)
  * @mixin \Eloquent
  */
 class UnionTypeModel extends Model